requirements.txt
fastapi
motor
passlib[bcrypt]
pydantic
python-dotenv
python-jose[cryptography]
python-multipart
uvicorn[standard]